requirements.txt
asgiref==3.2.3
Django==2.2.6
django-admin-sortable2==0.7.5
django-appconf==1.0.3
django-autoslug==1.9.6
django-ckeditor==5.8.0
django-constance==2.4.0
django-imagekit==4.0.2
django-js-asset==1.2.2
django-picklefield==2.0
django-redis==4.11.0
pilkit==2.0
Pillow==6.2.1
python-decouple==3.3
pytils==0.3
pytz==2019.3
redis==3.3.11
six==1.13.0
sqlparse==0.3.0
